+++ unirest-mocks/src/main/java/kong/unirest/Expectation.java
@@ -55,6 +55,13 @@ public interface Expectation {
     Expectation body(String body);
 
     /**
+     * A matcher for the body for a request
+     * @param matcher the matcher
+     * @return this Expectation
+     */
+    Expectation body(BodyMatcher matcher);
+
+    /**
      * expect a null response
      * @return The ExpectedResponse
      */
@@ -75,7 +82,7 @@ public interface Expectation {
     ExpectedResponse thenReturn(JSONElement jsonObject);
 
     /**
-     * expect a json response as defined by a pojo
+     * expect a object response as defined by a pojo using the requests / configuration object mapper
      * @param pojo the expected response body
      * @return The ExpectedResponse
      */