+++ unirest-mocks/src/main/java/kong/unirest/Invocation.java
@@ -26,24 +26,26 @@
 package kong.unirest;
 
 import kong.unirest.json.JSONElement;
-import org.apache.http.client.utils.URIBuilder;
 
-import java.net.URISyntaxException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import static java.lang.System.lineSeparator;
 
 class Invocation implements Expectation, ExpectedResponse {
     private Routes routes;
-    private String response;
+    private Function<ObjectMapper, String> response = o -> null;
     private Headers expectedHeaders = new Headers();
     private Headers expectedQueryParams = new Headers();
     private List<HttpRequest> requests = new ArrayList<>();
     private Headers responseHeaders = new Headers();
     private Boolean expected = false;
-    private String expectedBody;
+    private BodyMatcher expectedBody;
+    private MatchStatus expectedBodyStatus;
     private int responseStatus = 200;
     private String responseText = "Ok";
 
@@ -59,24 +61,30 @@ class Invocation implements Expectation, ExpectedResponse {
 
     @Override
     public ExpectedResponse thenReturn(String body) {
-        this.response = body;
+        this.response = o -> body;
         return this;
     }
 
     @Override
     public ExpectedResponse thenReturn(JSONElement jsonObject) {
-        this.response = jsonObject.toString();
+        this.response = o -> jsonObject.toString();
         return this;
     }
 
     @Override
     public ExpectedResponse thenReturn(Object pojo) {
-        this.response = new JsonObjectMapper().writeValue(pojo);
+        this.response = o -> o.writeValue(pojo);
         return this;
     }
 
-    RawResponse getResponse(Config config) {
-        return new MockRawResponse(response, responseHeaders, responseStatus, responseText, config);
+    RawResponse getResponse(Config config, HttpRequest request) {
+        return new MockRawResponse(response.apply(getObjectMapper(request, config)), responseHeaders, responseStatus, responseText, config);
+    }
+
+    private ObjectMapper getObjectMapper(HttpRequest request, Config config) {
+        return Util.tryCast(request, BaseRequest.class)
+                .map(BaseRequest::getObjectMapper)
+                .orElseGet(() -> config.getObjectMapper());
     }
 
     private Headers allHeaders() {
@@ -107,7 +115,13 @@ class Invocation implements Expectation, ExpectedResponse {
 
     @Override
     public Expectation body(String body) {
-        expectedBody = body;
+        expectedBody = new EqualsBodyMatcher(body);
+        return this;
+    }
+
+    @Override
+    public Expectation body(BodyMatcher matcher) {
+        expectedBody = matcher;
         return this;
     }
 
@@ -136,7 +150,7 @@ class Invocation implements Expectation, ExpectedResponse {
         }
         if(expectedBody != null){
             sb.append("Body:\n");
-            sb.append("\t" + expectedBody);
+            sb.append("\t" + expectedBodyStatus.getDescription());
         }
         return sb.toString();
     }
@@ -187,9 +201,7 @@ class Invocation implements Expectation, ExpectedResponse {
 
     private int scoreBody(HttpRequest request) {
         if(expectedBody != null){
-            return tryCast(request, HttpRequestUniBody.class)
-                    .map(HttpRequestUniBody::getBody)
-                    .flatMap(b -> b)
+            return tryCast(request, Body.class)
                     .map(this::matchBody)
                     .orElse(-1000);
         }
@@ -201,23 +213,30 @@ class Invocation implements Expectation, ExpectedResponse {
             BodyPart bodyPart = b.uniPart();
             if(bodyPart == null && expectedBody == null){
                 return 1;
-            } else if(String.class.isAssignableFrom(bodyPart.getPartType())
-            && expectedBody.equalsIgnoreCase((String) bodyPart.getValue())){
-                return 1;
+            } else if(String.class.isAssignableFrom(bodyPart.getPartType())){
+                expectedBodyStatus = expectedBody.matches(Arrays.asList((String) bodyPart.getValue()));
+                if(expectedBodyStatus.isSuccess()){
+                    return 1;
+                } else {
+                    return -1000;
+                }
             } else if (expectedBody != null) {
                 return -1000;
             }
+        } else {
+            List<String> parts = b.multiParts().stream().map(p -> p.toString()).collect(Collectors.toList());
+            if(parts.isEmpty() && expectedBody == null){
+                return 1;
+            } else {
+                expectedBodyStatus = expectedBody.matches(parts);
+                if(expectedBodyStatus.isSuccess()){
+                    return 1;
+                } else {
+                    return -1000;
+                }
+            }
         }
         return 0;
-    }
-
-    private URIBuilder geturi(HttpRequest request) {
-        try {
-            URIBuilder b = new URIBuilder(request.getUrl());
-            return b;
-        } catch (URISyntaxException e) {
-            throw new RuntimeException(e);
-        }
     }
 
     private int scoreHeaders(HttpRequest request) {
@@ -236,7 +255,7 @@ class Invocation implements Expectation, ExpectedResponse {
 
     private int scoreQuery(HttpRequest request) {
         if(expectedQueryParams.size() > 0){
-            URIBuilder p = geturi(request);
+            QueryParams p = QueryParams.fromURI(request.getUrl());
             long b = expectedQueryParams.all().stream().filter(h ->
                     p.getQueryParams().stream().anyMatch(q -> q.getName().equalsIgnoreCase(h.getName())
                             && q.getValue().equalsIgnoreCase(h.getValue())))