+++ unirest/src/test/java/BehaviorTests/CertificateTests.java
@@ -29,30 +29,12 @@ import kong.unirest.GetRequest;
 import kong.unirest.TestUtil;
 import kong.unirest.Unirest;
 import kong.unirest.UnirestException;
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.config.Registry;
-import org.apache.http.config.RegistryBuilder;
-import org.apache.http.conn.socket.ConnectionSocketFactory;
-import org.apache.http.conn.socket.PlainConnectionSocketFactory;
-import org.apache.http.conn.ssl.NoopHostnameVerifier;
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
-import org.apache.http.ssl.SSLContexts;
-import org.apache.http.util.EntityUtils;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLHandshakeException;
 import javax.net.ssl.SSLPeerUnverifiedException;
-
-import java.io.InputStream;
-import java.security.KeyStore;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -63,12 +45,13 @@ class CertificateTests extends BddTest {
 
     @Test
     void canDoClientCertificates() throws Exception {
-        Unirest.config().clientCertificateStore(readStore(), "badssl.com");
+        Unirest.config().clientCertificateStore(TestUtil.readStore(), "badssl.com");
 
         Unirest.get("https://client.badssl.com/")
                 .asString()
                 .ifFailure(r -> fail(r.getStatus() + " request failed " + r.getBody()))
-                .ifSuccess(r -> System.out.println(" woot "));;
+                .ifSuccess(r -> System.out.println(" woot "));
+        ;
     }
 
 
@@ -79,13 +62,14 @@ class CertificateTests extends BddTest {
         Unirest.get("https://client.badssl.com/")
                 .asString()
                 .ifFailure(r -> fail(r.getStatus() + " request failed " + r.getBody()))
-                .ifSuccess(r -> System.out.println(" woot "));;
+                .ifSuccess(r -> System.out.println(" woot "));
+        ;
     }
 
     @Test
     void loadWithSSLContext() throws Exception {
-        SSLContext sslContext = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadKeyMaterial(TestUtil.readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
                 .build();
 
         Unirest.config().sslContext(sslContext);
@@ -96,8 +80,8 @@ class CertificateTests extends BddTest {
 
     @Test
     void loadWithSSLContextAndProtocol() throws Exception {
-        SSLContext sslContext = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadKeyMaterial(TestUtil.readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
                 .build();
 
         Unirest.config().sslContext(sslContext).protocols("TLSv1.2");
@@ -108,8 +92,8 @@ class CertificateTests extends BddTest {
 
     @Test
     void loadWithSSLContextAndCipher() throws Exception {
-        SSLContext sslContext = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadKeyMaterial(TestUtil.readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
                 .build();
 
         Unirest.config().sslContext(sslContext).ciphers("TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384");
@@ -120,8 +104,8 @@ class CertificateTests extends BddTest {
 
     @Test
     void loadWithSSLContextAndCipherAndProtocol() throws Exception {
-        SSLContext sslContext = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadKeyMaterial(TestUtil.readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
                 .build();
 
         Unirest.config()
@@ -135,8 +119,8 @@ class CertificateTests extends BddTest {
 
     @Test
     void sslHandshakeFailsWhenServerIsReceivingAnUnsupportedCipher() throws Exception {
-        SSLContext sslContext = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadKeyMaterial(TestUtil.readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
                 .build();
 
         Unirest.config()
@@ -150,8 +134,8 @@ class CertificateTests extends BddTest {
 
     @Test
     void clientPreventsToUseUnsafeProtocol() throws Exception {
-        SSLContext sslContext = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
+        SSLContext sslContext = SSLContextBuilder.create()
+                .loadKeyMaterial(TestUtil.readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
                 .build();
 
         Unirest.config()
@@ -164,61 +148,10 @@ class CertificateTests extends BddTest {
 
     @Test
     void canSetHoestNameVerifyer() {
-        Unirest.config().hostnameVerifier(new NoopHostnameVerifier());
+        Unirest.config().hostnameVerifier((q, w) -> true);
 
         int response = Unirest.get("https://badssl.com/").asEmpty().getStatus();
         assertEquals(200, response);
-    }
-
-    @Test
-    void rawApacheClientCert() throws Exception {
-        SSLContext sslContext = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
-                .build();
-
-        HttpClient httpClient = HttpClients.custom().setSSLContext(sslContext).build();
-
-        HttpResponse response = httpClient.execute(new HttpGet("https://client.badssl.com/"));
-        assertEquals(200, response.getStatusLine().getStatusCode());
-        HttpEntity entity = response.getEntity();
-
-        System.out.println("----------------------------------------");
-        System.out.println(response.getStatusLine());
-        EntityUtils.consume(entity);
-    }
-
-    @Test
-    void rawApacheWithConnectionManager() throws Exception {
-        SSLContext sc = SSLContexts.custom()
-                .loadKeyMaterial(readStore(), "badssl.com".toCharArray()) // use null as second param if you don't have a separate key password
-                .build();
-
-        SSLConnectionSocketFactory sslSocketFactory =
-                new SSLConnectionSocketFactory(sc, new NoopHostnameVerifier());
-
-
-        Registry<ConnectionSocketFactory> socketFactoryRegistry =
-                RegistryBuilder.<ConnectionSocketFactory>create()
-                        .register("https", sslSocketFactory)
-                        .register("http", PlainConnectionSocketFactory.INSTANCE)
-                        .build();
-
-        PoolingHttpClientConnectionManager cm =
-                new PoolingHttpClientConnectionManager(socketFactoryRegistry);
-
-        CloseableHttpClient httpClient =
-                HttpClients.custom()
-                        .setSSLSocketFactory(sslSocketFactory)
-                        .setConnectionManager(cm)
-                        .build();
-
-        Unirest.config().httpClient(httpClient);
-
-        Unirest.get("https://client.badssl.com/")
-                .asString()
-                .ifFailure(r -> fail(r.getStatus() + " request failed " + r.getBody()))
-                .ifSuccess(r -> System.out.println(" woot "));
-
     }
 
     @Test
@@ -236,8 +169,8 @@ class CertificateTests extends BddTest {
     void expired() {
         fails("https://expired.badssl.com/",
                 SSLHandshakeException.class,
-                "javax.net.ssl.SSLHandshakeException: sun.security.validator.ValidatorException: " +
-                        "PKIX path validation failed: java.security.cert.CertPathValidatorException: validity check failed");
+                "javax.net.ssl.SSLHandshakeException: PKIX path validation failed: " +
+                        "java.security.cert.CertPathValidatorException: validity check failed");
         disableSsl();
         canCall("https://expired.badssl.com/");
     }
@@ -246,7 +179,7 @@ class CertificateTests extends BddTest {
     void selfSigned() {
         fails("https://self-signed.badssl.com/",
                 SSLHandshakeException.class,
-                "javax.net.ssl.SSLHandshakeException: sun.security.validator.ValidatorException: " +
+                "javax.net.ssl.SSLHandshakeException: " +
                         "PKIX path building failed: sun.security.provider.certpath.SunCertPathBuilderException: " +
                         "unable to find valid certification path to requested target");
         disableSsl();
@@ -268,7 +201,7 @@ class CertificateTests extends BddTest {
     void expiredAsync() {
         failsAsync("https://expired.badssl.com/",
                 SSLHandshakeException.class,
-                "javax.net.ssl.SSLHandshakeException: General SSLEngine problem");
+                "javax.net.ssl.SSLHandshakeException: PKIX path validation failed: java.security.cert.CertPathValidatorException: validity check failed");
         disableSsl();
         canCallAsync("https://expired.badssl.com/");
     }
@@ -277,7 +210,8 @@ class CertificateTests extends BddTest {
     void selfSignedAsync() {
         failsAsync("https://self-signed.badssl.com/",
                 SSLHandshakeException.class,
-                "javax.net.ssl.SSLHandshakeException: General SSLEngine problem");
+                "javax.net.ssl.SSLHandshakeException: PKIX path building failed: sun.security.provider.certpath.SunCertPathBuilderException: " +
+                        "unable to find valid certification path to requested target");
         disableSsl();
         canCallAsync("https://self-signed.badssl.com/");
     }
@@ -310,11 +244,4 @@ class CertificateTests extends BddTest {
         }
     }
 
-    private KeyStore readStore() throws Exception {
-        try (InputStream keyStoreStream = this.getClass().getResourceAsStream("/certs/badssl.com-client.p12")) {
-            KeyStore keyStore = KeyStore.getInstance("PKCS12");
-            keyStore.load(keyStoreStream, "badssl.com".toCharArray());
-            return keyStore;
-        }
-    }
 }