+++ unirest/src/test/java/BehaviorTests/InterceptorTest.java
@@ -26,15 +26,9 @@
 package BehaviorTests;
 
 import kong.unirest.*;
-import org.apache.http.HttpException;
-import org.apache.http.HttpHost;
-import org.apache.http.HttpRequestInterceptor;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpUriRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
@@ -42,14 +36,11 @@ import java.util.concurrent.ExecutionException;
 import static com.google.common.collect.Sets.newHashSet;
 import static kong.unirest.TestUtil.rezFile;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 class InterceptorTest extends BddTest {
 
     private UniInterceptor interceptor;
-    private final String ioErrorMessage = "Something horrible happened";;
+
 
     @BeforeEach
     public void setUp() {
@@ -89,74 +80,35 @@ class InterceptorTest extends BddTest {
 
     @Test
     void totalFailure() throws Exception {
-        Unirest.config().httpClient(getFailureClient()).interceptor(interceptor);
+        Unirest.config().httpClient(TestUtil.getFailureClient()).interceptor(interceptor);
 
         TestUtil.assertException(() -> Unirest.get(MockServer.GET).asEmpty(),
                 UnirestException.class,
-                "java.io.IOException: " + ioErrorMessage);
+                "java.io.IOException: " + "Something horrible happened");
     }
 
     @Test
     void canReturnEmptyResultRatherThanThrow() throws Exception {
-        Unirest.config().httpClient(getFailureClient()).interceptor(interceptor);
+        Unirest.config().httpClient(TestUtil.getFailureClient()).interceptor(interceptor);
         interceptor.failResponse = true;
 
         HttpResponse<String> response = Unirest.get(MockServer.GET).asString();
 
         assertEquals(542, response.getStatus());
-        assertEquals(ioErrorMessage, response.getStatusText());
-    }
-
-    @Test
-    void totalAsyncFailure() throws Exception {
-        Unirest.config().addInterceptor((r, c) -> {
-            throw new IOException(ioErrorMessage);
-        }).interceptor(interceptor);
-
-        TestUtil.assertException(() -> Unirest.get(MockServer.GET).asStringAsync().get(),
-                ExecutionException.class,
-                "java.io.IOException: " + ioErrorMessage);
+        assertEquals("Something horrible happened", response.getStatusText());
     }
 
     @Test
     void totalAsyncFailure_Recovery() throws Exception {
         interceptor.failResponse = true;
-        Unirest.config().addInterceptor((r, c) -> {
-            throw new IOException(ioErrorMessage);
-        }).interceptor(interceptor);
+        Unirest.config()
+                .asyncClient(TestUtil.getFailureAsyncClient())
+                .interceptor(interceptor);
 
         HttpResponse<String> response = Unirest.get(MockServer.GET).asStringAsync().get();
 
         assertEquals(542, response.getStatus());
-        assertEquals(ioErrorMessage, response.getStatusText());
-    }
-
-    private HttpClient getFailureClient() throws IOException {
-        HttpClient client = mock(HttpClient.class);
-        when(client.execute(any(HttpHost.class), any(HttpUriRequest.class))).thenThrow(new IOException(ioErrorMessage));
-        when(client.execute(any(HttpUriRequest.class))).thenThrow(new IOException(ioErrorMessage));
-        return client;
-    }
-
-    @Test @Deprecated
-    void canAddApacheInterceptor() {
-        Unirest.config().addInterceptor(new TestInterceptor());
-
-        Unirest.get(MockServer.GET)
-                .asObject(RequestCapture.class)
-                .getBody()
-                .assertHeader("x-custom", "foo");
-    }
-
-    @Test @Deprecated
-    void canAddApacheInterceptorToAsync() throws ExecutionException, InterruptedException {
-        Unirest.config().addInterceptor(new TestInterceptor());
-
-        Unirest.get(MockServer.GET)
-                .asObjectAsync(RequestCapture.class)
-                .get()
-                .getBody()
-                .assertHeader("x-custom", "foo");
+        assertEquals("Something horrible happened", response.getStatusText());
     }
 
     @Test
@@ -179,12 +131,7 @@ class InterceptorTest extends BddTest {
         assertEquals(newHashSet("file=spidey.jpg","fruit=apples"), values);
     }
 
-    private class TestInterceptor implements HttpRequestInterceptor {
-        @Override
-        public void process(org.apache.http.HttpRequest httpRequest, org.apache.http.protocol.HttpContext httpContext) throws HttpException, IOException {
-            httpRequest.addHeader("x-custom", "foo");
-        }
-    }
+
 
     private class UniInterceptor implements Interceptor {
         RequestCapture cap;