+++ unirest/src/test/java/kong/unirest/ClientFactoryTest.java
@@ -25,20 +25,13 @@
 
 package kong.unirest;
 
-import org.apache.http.HttpRequestInterceptor;
-import org.apache.http.impl.nio.client.CloseableHttpAsyncClient;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.lang.management.ManagementFactory;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.lessThan;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.mockito.Mockito.mock;
+import static org.junit.jupiter.api.Assertions.*;
 
 class ClientFactoryTest {
 
@@ -52,24 +45,7 @@ class ClientFactoryTest {
     void shouldReuseThreadPool() {
         int startingCount = ManagementFactory.getThreadMXBean().getThreadCount();
         //IntStream.range(0,100).forEach(i -> ClientFactory.refresh());
-        assertThat(ManagementFactory.getThreadMXBean().getThreadCount(), is(lessThan(startingCount + 10)));
-    }
-
-    @Test
-    void canSaveSomeOptions(){
-        HttpRequestInterceptor i = mock(HttpRequestInterceptor.class);
-        CloseableHttpAsyncClient c = mock(CloseableHttpAsyncClient.class);
-
-        Unirest.config()
-                .addInterceptor(i)
-                .connectTimeout(4000)
-                .asyncClient(c);
-
-        Unirest.shutDown(false);
-
-        assertNotEquals(c, Unirest.config().getAsyncClient());
-        assertEquals(i, Unirest.config().getInterceptor().get(0));
-        assertEquals(4000, Unirest.config().getConnectionTimeout());
+        assertTrue(ManagementFactory.getThreadMXBean().getThreadCount() < startingCount + 10);
     }
 
 }