+++ unirest/src/test/java/kong/unirest/TestUtil.java
@@ -26,15 +26,23 @@
 package kong.unirest;
 
 import BehaviorTests.MockServer;
-import BehaviorTests.UploadProgressTest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.guava.GuavaModule;
 import com.google.common.base.Charsets;
 import com.google.common.base.Throwables;
 import com.google.common.io.Resources;
+import kong.unirest.apache.ApacheAsyncClient;
+import kong.unirest.apache.ApacheClient;
+import org.apache.http.HttpHost;
+import org.apache.http.HttpRequestInterceptor;
+import org.apache.http.client.HttpClient;
+import org.apache.http.client.methods.HttpUriRequest;
+import org.apache.http.impl.nio.client.CloseableHttpAsyncClient;
+import org.apache.http.impl.nio.client.HttpAsyncClientBuilder;
 
 import java.io.*;
 import java.net.URISyntaxException;
+import java.security.KeyStore;
 import java.time.Instant;
 import java.util.Base64;
 import java.util.HashMap;
@@ -45,6 +53,9 @@ import java.util.stream.Collectors;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.fail;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class TestUtil {
     private static final ObjectMapper om = new ObjectMapper();
@@ -98,14 +109,6 @@ public class TestUtil {
         }
     }
 
-    public static InputStream emptyInput() {
-        return new ByteArrayInputStream(new byte[]{});
-    }
-
-    public static InputStream toInputStream(String s) {
-        return new ByteArrayInputStream(s.getBytes());
-    }
-
     public static <K, V> Map<K, V> mapOf(Object... keyValues) {
         Map<K, V> map = new HashMap<>();
 
@@ -120,21 +123,6 @@ public class TestUtil {
         }
 
         return map;
-    }
-
-    public static void debugApache() {
-        System.setProperty("org.apache.commons.logging.Log","org.apache.commons.logging.impl.SimpleLog");
-        System.setProperty("org.apache.commons.logging.simplelog.showdatetime", "true");
-        System.setProperty("org.apache.commons.logging.simplelog.log.org.apache.http", "DEBUG");
-        System.setProperty("org.apache.commons.logging.simplelog.log.org.apache.http.wire", "DEBUG");
-    }
-
-    public static <T> T read(String o, Class<T> as){
-        try {
-            return om.readValue(o, as);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
     }
 
     public static void assertBasicAuth(String raw, String username, String password) {
@@ -188,6 +176,30 @@ public class TestUtil {
 
     public static void reset() {
         Util.resetClock();
+    }
+
+    public static KeyStore readStore() throws Exception {
+        try (InputStream keyStoreStream = TestUtil.class.getResourceAsStream("/certs/badssl.com-client.p12")) {
+            KeyStore keyStore = KeyStore.getInstance("PKCS12");
+            keyStore.load(keyStoreStream, "badssl.com".toCharArray());
+            return keyStore;
+        }
+    }
+
+    public static Client getFailureClient() throws IOException {
+        HttpClient client = mock(HttpClient.class);
+        when(client.execute(any(HttpHost.class), any(HttpUriRequest.class))).thenThrow(new IOException("Something horrible happened"));
+        when(client.execute(any(HttpUriRequest.class))).thenThrow(new IOException("Something horrible happened"));
+        return new ApacheClient(client, Unirest.config());
+    }
+
+    public static AsyncClient getFailureAsyncClient() {
+        CloseableHttpAsyncClient client = HttpAsyncClientBuilder.create()
+                .addInterceptorFirst((HttpRequestInterceptor) (r, c) -> {
+            throw new IOException("Something horrible happened");
+        }).build();
+        client.start();
+        return new ApacheAsyncClient(client, Unirest.config());
     }
 
     @FunctionalInterface