+++ CHANGELOG.md
@@ -1,4 +1,22 @@
-## 3.11.06 (pending)
+## 3.11.11 (pending)
+* useSystemProperties is not overwritten and should be reset back to true on reset
+
+## 3.11.10
+* issue #394 use the configured Object Mapper rather than always Json
+* internal pre-factorings to get ready for Unirest 4
+
+## 3.11.09
+* Expected body param values for Mock expects need to be url encoded
+* Support ANY expectation on methods for MockClient. (e.g. ```expect(HttpMethod.GET)``)
+
+## 3.11.08
+* Adds new body matchers to the Mock client for asserting multipart forms.
+
+## 3.11.07
+* add a new object constructor on JSONObject for basic object serialization
+* issue #392 overwrite non-specified number serialization
+  
+## 3.11.06
 * add a convenience method for setting the content type
 * add a common reference to popular mime types
 * cache methods on Config were not returning the config for the builder pattern.